Prevent breaking of frames to prevent page break bug
When using mdframed frames within columns, there is some weird
behaviour that frames get broken up unnecessarily, and then a page
break is inserted also unneccessarily. Strange behaviour.

Since we only use small frames in the 'infobox' env, it is no
problem and actually better for the design to not break these.